requirements.txt
# Core Dependencies
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
scipy==1.11.3
scikit-learn==1.3.0

# Deep Learning and AI
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
mtcnn==0.1.1
facenet-pytorch==2.5.3

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
alembic==1.12.0
redis==4.6.0

# Web Framework and API
FastAPI==0.103.2
uvicorn==0.23.2
starlette==0.27.0
pydantic==2.4.2
jinja2==3.1.2
aiofiles==23.2.1

# Dashboard
streamlit==1.29.0
plotly==5.17.0
websockets==12.0

# Image Processing and Computer Vision
dlib==19.24.2
face-recognition==1.3.0
imageio==2.31.3
scikit-image==0.21.0

# Utilities and Configuration
PyYAML==6.0.1
python-dotenv==1.0.0
click==8.1.7
rich==13.5.2
tqdm==4.66.1

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.17.1
grafana-api==1.0.3

# Communication and Alerts
requests==2.31.0
smtplib-ssl==1.0.4
twilio==8.5.0
slack-sdk==3.22.0

# Security and Authentication
cryptography==41.0.4
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
argon2-cffi==23.1.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
httpx==0.25.0

# Development Tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1
pre-commit==3.4.0

# Performance and Optimization
numba==0.57.1
faiss-cpu==1.7.4  # Use faiss-gpu if GPU available
joblib==1.3.2

# Video Processing
ffmpeg-python==0.2.0
imageio-ffmpeg==0.4.9

# Data Analysis and Visualization
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Async and Concurrency
aiohttp==3.8.5
aioredis==2.0.1
celery==5.3.2

# Liveness Detection
scipy==1.11.3
timm==0.9.2  # For MiDaS depth estimation

# Clustering
hdbscan==0.8.33

# Edge Deployment
onnx==1.15.0
onnxruntime==1.16.0  # Use onnxruntime-gpu for GPU

# Mobile Integration
PyJWT==2.8.0
google-auth==2.23.0

# Demographics
# Models use torch (already included)

# Re-ID (Multi-camera tracking)
torchreid==0.2.5

# GPU Support (Optional - uncomment if using GPU)
# tensorflow-gpu==2.13.0
# faiss-gpu==1.7.4

# Development Dependencies
# jupyter==1.0.0
# notebook==7.0.3
# ipykernel==6.25.2